[+] add UpdatePictureWorker worker
